AccountManagerPlugin: Handle empty password files without exception, refs #8563.
AccountManagerPlugin terminates each line, so there's no such thing
like an empty file because at least one linebreak is always retained.
But other application may not equally care for proper line endings.
So I've moved the existing check for data read length up
to prevent reading from a possibly empty list in case of a blank file.
Checking EOL style not in first but in the last line is the other change,
but certainly unrelated to the issue.