Correct end-of-day adjustments for ASAP and < 100% effort.  Refs #9533, #9300.
When working forward to build an ASAP schedule, partial takes caused the
tasks to shift forward more than their scheduled length.

If one resource can work 8 hours per day and another can only work 4,
the second resource takes twice as long to complete the same task. But
without this clean up, ALAP scheduling took an 8-hour task at 4 hours
per day from the end of day 2 to the middle of day 1.  With this change,
if all available hours are used up, the rest of the day is consumed so
that 8-hour task goes from the end of day 2 to the start of day 1.