AccountManagerPlugin: Add flexible date/time rendering for user lists, refs #6616 and #9852.
Now the time stamps are combined with a relative time interval hint (tool-tip).
This is an enhancement to the user admin panel too.

Support for bleeding-edge user configurable time in Trac 0.13 is accompanied
here by a fallback for Trac 0.11 and 0.12, that looks great and is worth a lot
of the effort put into this rather complicated fallback code.