Make group tickets depend on their milestones.  Refs #10313.
This is an imperfect fix.  I don't like that the chart shows an explicit
dependency between the group and the milestone.  But it does make the
schedule correct.

I'd like to undo this and add code in the area where I propagate
dependencies down from parents to children to make groups depend on
their milestone only in the augmented graph within the scheduler.  That
would make the schedule correct and eliminate the stray dependency line
on the chart.  But that's more work than I can do now and this is
better.