Complete rework of plugin, using ideas from #11210 in the patch by chris-at. Fixes #7554, Refs #8821, #11210.
Note that only one instance of the reporter in `to_recipients` is removed. So for instance, if the reporter is also the owner and `always_notify_owner` is `True`, then the reporter will still receive a notification.

This has been lightly tested, so please report back if you find any issues.